+++ src/main/scala/de/upb/cs/uc4/hyperledger/connections/traits/ConnectionTrait.scala
@@ -1,34 +1,39 @@
 package de.upb.cs.uc4.hyperledger.connections.traits
 
 import java.nio.charset.StandardCharsets
+import java.nio.file.Path
 import java.util
-import java.util.Optional
 import java.util.concurrent.TimeoutException
 
 import com.google.protobuf.ByteString
+import de.upb.cs.uc4.hyperledger.connections.cases.ConnectionApproval
 import de.upb.cs.uc4.hyperledger.exceptions.traits.{ HyperledgerExceptionTrait, TransactionExceptionTrait }
 import de.upb.cs.uc4.hyperledger.exceptions.{ HyperledgerException, NetworkException, TransactionException }
-import de.upb.cs.uc4.hyperledger.utilities.helper.{ ReflectionHelper, TransactionHelper }
+import de.upb.cs.uc4.hyperledger.utilities.ConnectionManager
+import de.upb.cs.uc4.hyperledger.utilities.helper.{ Logger, ReflectionHelper, TransactionHelper }
 import org.hyperledger.fabric.gateway.impl.{ ContractImpl, GatewayImpl, TransactionImpl }
 import org.hyperledger.fabric.gateway.GatewayRuntimeException
-import org.hyperledger.fabric.protos.peer.{ Chaincode, ProposalPackage }
+import org.hyperledger.fabric.protos.peer.ProposalPackage
 import org.hyperledger.fabric.protos.peer.ProposalPackage.{ Proposal, SignedProposal }
 import org.hyperledger.fabric.sdk._
 import org.hyperledger.fabric.sdk.transaction.{ ProposalBuilder, TransactionContext }
 
-import scala.collection.convert.ImplicitConversions.`iterator asScala`
 import scala.jdk.CollectionConverters.MapHasAsJava
 
 trait ConnectionTrait extends AutoCloseable {
-  // regular contract info
-  val contractName: String
-  val contract: ContractImpl
-  val gateway: GatewayImpl
+  // regular info used to set up any connection
+  val username: String
+  val channel: String
+  val chaincode: String
+  val walletPath: Path
+  val networkDescriptionPath: Path
 
-  // draftContract info
-  val draftContractName: String = "UC4.Draft"
-  val draftContract: ContractImpl
-  val draftGateway: GatewayImpl
+  // contract info for specific connections
+  val contractName: String
+
+  // setting up connections
+  lazy val (contract: ContractImpl, gateway: GatewayImpl) = ConnectionManager.initializeConnection(username, channel, chaincode, contractName, walletPath, networkDescriptionPath)
+  lazy val approvalConnection: Option[ConnectionApprovalsTrait] = Some(ConnectionApproval(username, channel, chaincode, walletPath, networkDescriptionPath))
 
   /** Gets the version returned by the designated contract.
     * By default all contracts return the version of the chaincode.
@@ -40,47 +45,90 @@ trait ConnectionTrait extends AutoCloseable {
   /** Wrapper for a submission transaction
     * Translates the result byte-array to a string and throws an error if said string contains an error.
     *
-    * @param transient     boolean flag to determine transaction to be transient or not.
-    * @param transactionId transaction to call
-    * @param params        parameters to feed into transaction
+    * @param transient        boolean flag to determine transaction to be transient or not.
+    * @param transactionName  transaction to call
+    * @param params           parameters to feed into transaction
     * @return result as a string
     */
   @throws[TransactionExceptionTrait]
   @throws[HyperledgerExceptionTrait]
-  protected final def wrapSubmitTransaction(transient: Boolean, transactionId: String, params: String*): String = {
-    val result = this.privateSubmitTransaction(transient, transactionId, params: _*)
-    this.wrapTransactionResult(transactionId, result)
+  protected final def wrapSubmitTransaction(transient: Boolean, transactionName: String, params: String*): String = {
+    // submit my approval to approvalContract
+    if (approvalConnection.isDefined) approvalConnection.get.approveTransaction(contractName, transactionName, params: _*)
+
+    // submit and evaluate response from my "regular" contract
+    val result = this.privateSubmitTransaction(transient, transactionName, params: _*)
+    this.wrapTransactionResult(transactionName, result)
   }
 
   /** Wrapper for an evaluation transaction
     * Translates the result byte-array to a string and throws an error if said string contains an error.
     *
-    * @param transactionId transaction to call
+    * @param transactionName transaction to call
     * @param params        parameters to feed into transaction
     * @return result as a string
     */
   @throws[TransactionExceptionTrait]
   @throws[HyperledgerExceptionTrait]
-  protected final def wrapEvaluateTransaction(transactionId: String, params: String*): String = {
-    val result = this.privateEvaluateTransaction(transactionId, params: _*)
-    this.wrapTransactionResult(transactionId, result)
+  protected final def wrapEvaluateTransaction(transactionName: String, params: String*): String = {
+    // submit my approval to approvalContract
+    if (approvalConnection.isDefined) approvalConnection.get.approveTransaction(contractName, transactionName, params: _*)
+
+    val result = this.privateEvaluateTransaction(transactionName, params: _*)
+    this.wrapTransactionResult(transactionName, result)
   }
 
   protected final def internalGetUnsignedProposal(transactionName: String, params: String*): Array[Byte] = {
-    val (transaction, context, request) = TransactionHelper.createTransactionInfo(contract, transactionName, params.toArray, None)
+    // send approval as admin maintaining the connection
+    approvalConnection.get.approveTransaction(contractName, "addCertificate", params: _*)
 
-    val proposalBuilder: ProposalBuilder = ProposalBuilder.newBuilder()
-      .request(request)
-      .context(context)
-    val proposal: Proposal = proposalBuilder.build()
+    // gather info
+    val (transaction, context, request) = TransactionHelper.createApprovalTransactionInfo(approvalConnection.get.contract, contractName, transactionName, params.toArray, None)
+
+    // create proposal
+    val proposal: Proposal = ProposalBuilder.newBuilder().request(request).context(context).build()
     proposal.toByteArray
   }
 
-  def submitSignedProposal(proposalBytes: Array[Byte], signature: ByteString): String = {
+  def submitSignedProposal(proposalBytes: Array[Byte], signatureBytes: Array[Byte]): String = {
     val proposal: Proposal = Proposal.parseFrom(proposalBytes)
+    val signature: ByteString = ByteString.copyFrom(signatureBytes)
 
-    val (transaction, context, signedProposal) = this.createProposal(proposal, signature)
+    val (transaction: TransactionImpl, context: TransactionContext, signedProposal: SignedProposal) = this.createSignedProposal(proposal, signature)
+
+    // submit approval
+    val approvalResult = this.internalSubmitApprovalProposal(transaction, context, signedProposal)
+    var transactionResult = approvalResult
+    try {
+      // submit real transaction as admin
+      transactionResult = this.internalSubmitRealTransactionFromApprovalProposal(proposal)
+    }
+    catch {
+      case e: TransactionExceptionTrait => throw e
+    }
+    transactionResult
+  }
+
+  def internalSubmitRealTransactionFromApprovalProposal(proposal: Proposal): String = {
+    val (proposalContractName, transactionName, params) = TransactionHelper.getParametersFromApprovalProposal(proposal)
+
+    // Logging
+    Logger.warn("contractName" + proposalContractName)
+    Logger.warn("transactionName" + transactionName)
+    Logger.warn("params" + params.mkString(";"))
+
+    // check contract match
+    if (proposalContractName != contractName) throw TransactionException.CreateUnknownException("approveTransaction", s"Approval was sent to wrong connection:: $contractName != $proposalContractName")
+
+    // submit and evaluate response from my "regular" contract
+    // TODO: pass transient bool
+    val result = this.privateSubmitTransaction(false, transactionName, params.toIndexedSeq: _*)
+    this.wrapTransactionResult(transactionName, result)
+  }
+
+  def internalSubmitApprovalProposal(transaction: TransactionImpl, context: TransactionContext, signedProposal: SignedProposal): String = {
     val transactionName = TransactionHelper.getTransactionNameFromFcn(transaction.getName)
+    if (transactionName != "approveTransaction") throw new Exception("submitSigned Proposal was invoked with a non approval transaction.")
     val proposalResponses = this.sendProposalToPeers(context, signedProposal)
 
     // evaluate proposal
@@ -94,7 +142,7 @@ trait ConnectionTrait extends AutoCloseable {
     }
   }
 
-  private def createProposal(proposal: ProposalPackage.Proposal, signature: ByteString): (TransactionImpl, TransactionContext, SignedProposal) = {
+  private def createSignedProposal(proposal: ProposalPackage.Proposal, signature: ByteString): (TransactionImpl, TransactionContext, SignedProposal) = {
     val transactionId: String = TransactionHelper.getTransactionIdFromProposal(proposal)
     val transactionName: String = TransactionHelper.getTransactionNameFromProposal(proposal)
     val params: Seq[String] = TransactionHelper.getTransactionParamsFromProposal(proposal)
@@ -104,7 +152,7 @@ trait ConnectionTrait extends AutoCloseable {
       .setSignature(signature)
     val signedProposal: SignedProposal = signedProposalBuilder.build
 
-    val (transaction, context, request) = TransactionHelper.createTransactionInfo(contract, transactionName, params.toArray, Some(transactionId))
+    val (transaction, context, request) = TransactionHelper.createTransactionInfo(approvalConnection.get.contract, transactionName, params.toArray, Some(transactionId))
 
     (transaction, context, signedProposal)
   }
@@ -117,15 +165,20 @@ trait ConnectionTrait extends AutoCloseable {
 
   @throws[HyperledgerExceptionTrait]
   private def privateSubmitTransaction(transient: Boolean, transactionName: String, params: String*): Array[Byte] = {
-    testParamsNull(transactionName, params: _*)
+    testAnyParamsNull(transactionName, params: _*)
     try {
       if (transient) {
-        var transMap: Map[String, Array[Byte]] = Map()
         var i = 0
+        var transMap: Map[String, Array[Byte]] = Map()
         params.foreach(param => {
           transMap += i.toString -> param.toCharArray.map(_.toByte)
           i = i + 1
         })
+
+        // TODO: once transient is reenabled, test this
+        // var i = 0
+        // val transMap2 = params.toList.map((entry: String) =>
+        // ((i+=1).toString -> entry.toCharArray.map(_.toByte)))
 
         contract.createTransaction(transactionName).setTransient(transMap.asJava).submit()
       }
@@ -142,7 +195,7 @@ trait ConnectionTrait extends AutoCloseable {
 
   @throws[HyperledgerExceptionTrait]
   private def privateEvaluateTransaction(transactionName: String, params: String*): Array[Byte] = {
-    testParamsNull(transactionName, params: _*)
+    testAnyParamsNull(transactionName, params: _*)
     try {
       contract.evaluateTransaction(transactionName, params: _*)
     }
@@ -182,7 +235,7 @@ trait ConnectionTrait extends AutoCloseable {
     * @throws TransactionException if a parameter is null.
     */
   @throws[TransactionExceptionTrait]
-  private def testParamsNull(transactionName: String, params: String*): Unit = {
-    params.foreach(param => if (param == null) throw TransactionException.CreateUnknownException(transactionName, "A parameter was null."))
+  private def testAnyParamsNull(transactionName: String, params: String*): Unit = {
+    if (params.exists(a => a == null)) throw TransactionException.CreateUnknownException(transactionName, "A parameter was null.")
   }
 }